backend/requirements.txt
asgiref==3.3.0
Django==3.1.3
pytz==2020.4
sqlparse==0.4.1
djangorestframework==3.12.1
django_cors_headers==3.5.0